# Optimize parser performance