**Detailed description**:
- Fix slow string creation and comparison in transaction filter by using EntityId
- Fix slow metric recording by pre-creating metrics objects in constructor
- Fix slow configuration properties getters due to reflection by removing `@Validated`
- Fix saving record file before all batches complete (recordfile save is in separate transaction)
- Fix NullPointerException if exception during `StreamFileListener.onStart()`
- Fix slow file reading due to unbuffered input stream
- Fix unnecessary ByteBuffer allocation by using library to convert int to bytes
- Fix unnecessary String creation for transaction memo

**Which issue(s) this PR fixes**:

**Special notes for your reviewer**:

These statistics from running the importer locally on my Macbook. As such, only the relative performance gain from before and after can be considered.

**RecordFileParserPerformanceTest**:
  Before: ~6s
  After: ~3s

**Perf test bucket with test run of 3500TPS at 256B topic messages closing every 5s**:
  Database state: 20M transactions, 20M topic messages and 61M cryptotransfers
  Before: ~4.05s per record file
  After: ~1.827s perf record file

**Perf test bucket with test run of 3500TPS at 256B topic messages closing every 1s**:
  Database state: 16M transactions, 16M topic messages and 48M cryptotransfers in DB
  Before: ~1.2975s per record file
  After: ~.396s avg per record file

So anywhere from a 2x to a 3x improvement

**Checklist**
- [ ] Documentation added
- [ ] Tests updated